DEV-TOOLS: Fix data segment seg references into rtlink segments
For Xanth, the data segment has segment references into rtlink segment
areas. Since this couldn't be narrowed down to one specific segment,
the reference was just random, and worse, it was splitting up some
of the rtlink segments the references fell on in the decoded Exe.
It was simplest just to remove the relocation entries for such
references.